Integration test minor fix move decodeProtobufPayload to commons-test
