+++ packages/catalog-process/package.json
@@ -48,6 +48,7 @@
     "mongodb": "5.6.0",
     "openapi-zod-client": "1.15.1",
     "pagopa-interop-commons": "workspace:*",
+    "pagopa-interop-commons-test": "workspace:*",
     "pagopa-interop-models": "workspace:*",
     "ts-pattern": "^5.0.6",
     "uuid": "^9.0.0",
