+++ packages/commons-test/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@anatine/zod-mock": "^3.13.1",
     "@pagopa/eslint-config": "^3.0.0",
+    "@protobuf-ts/runtime": "^2.9.1",
     "@types/uuid": "^9.0.2",
     "@zodios/core": "^10.9.2",
     "pagopa-interop-commons": "workspace:*",
